requirements.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ML/Computer Vision
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
opencv-python>=4.8.0

# PDF Generation
reportlab>=4.0.0
fpdf2>=2.7.0

# External APIs
httpx>=0.25.0
openai>=1.0.0  # For GPT summary (optional)

# Utilities
python-dotenv>=1.0.0

